==> 2_base_de_dados.ipynb <==
## Base de dados

É necessário utilizar os arquivos gerados na fase de Scraping.


import numpy as np
import pandas as pd
import re

A primeira tarefa é cadastrar os times numa base de dados.

Vamos considerar todos os times que disputaram ao menos uma temporada na primeira divisão do seu país desde 2002.

#MontarBase.py

#lê os .csv e gera um dataframe para cada país
df_arg = pd.read_csv("data/Argentina.csv")
df_bol = pd.read_csv("data/Bolivia.csv")
df_bra = pd.read_csv("data/Brasil.csv")
df_chi = pd.read_csv("data/Chile.csv")
df_col = pd.read_csv("data/Colombia.csv")
df_ecu = pd.read_csv("data/Equador.csv")
df_par = pd.read_csv("data/Paraguai.csv")
df_per = pd.read_csv("data/Peru.csv")
df_uru = pd.read_csv("data/Uruguai.csv")
df_ven = pd.read_csv("data/Venezuela.csv")
df_arg['pais'] = "Argentina"
df_bol['pais'] = "Bolivia"
df_bra['pais'] = "Brasil"
df_chi['pais'] = "Chile"
df_col['pais'] = "Colombia"
df_ecu['pais'] = "Equador"
df_par['pais'] = "Paraguai"
df_per['pais'] = "Peru"
df_uru['pais'] = "Uruguai"
df_ven['pais'] = "Venezuela"

dfs = [df_arg,df_bol,df_bra,df_chi,df_col,df_ecu,df_par,df_per,df_uru,df_ven]
        
for df in dfs:
    dates = []
    for i in range(df['date'].size):        
        item = df['date'].get(i)
        date = item.split('.')
        if df['pais'][0] == 'Argentina' or df['pais'][0] == 'Venezuela':
            if date[1]>'06':
                date[2]=date[2][0:4]
            else:
                date[2]=date[2][-4:]
        item = date[2]+'.'+date[1]+'.'+date[0]
        dates.append(item)
    df['date']=dates
    df.sort_values(by="date", inplace=True)
    df = df.drop(columns = 'pais')

#cria o cadastro de times argentinos
df_teste = df_arg[df_arg['date']>='2002.01.01'][df_arg['tournament']>='Primera Division'][df_arg['tournament']!='Super Cup']
clubs = df_teste['club_1'].values, df_teste['club_2'].values
clubs = np.unique(clubs)
clubes=[]
for i in range(len(clubs)):
    clubes.append([i+1, clubs[i], 'Argentina',1500])
df_clubarg = pd.DataFrame(clubes)

#cria o cadastro de times bolivianos
df_teste = df_bol[df_bol['date']>='2002.01.01']
clubs = df_teste['club_1'].values, df_teste['club_2'].values
clubs = np.unique(clubs)
clubes=[]
for i in range(len(clubs)):
    clubes.append([i+101, clubs[i], 'Bolivia',1500])
df_clubbol = pd.DataFrame(clubes)

#cria o cadastro de times brasileiros 
df_teste = df_bra[df_bra['date']>='2002.01.01'][df_bra['tournament']=='Série A']
clubs = df_teste['club_1'].values, df_teste['club_2'].values
clubs = np.unique(clubs)
clubes=[]
for i in range(len(clubs)):
    clubes.append([i+201, clubs[i], 'Brasil',1500])
df_clubbra = pd.DataFrame(clubes)

#cria o cadastro de times chilenos 
df_teste = df_chi[df_chi['date']>='2002.01.01'][df_chi['tournament']=='Primera Division']
clubs = df_teste['club_1'].values, df_teste['club_2'].values
clubs = np.unique(clubs)
clubes=[]
for i in range(len(clubs)):
    clubes.append([i+301, clubs[i], 'Chile',1500])
df_clubchi = pd.DataFrame(clubes)

#cria o cadastro de times colombianos 
df_teste = df_col[df_col['date']>='2002.01.01'][df_col['tournament']>='Liga'][df_col['tournament']<'M']
clubs = df_teste['club_1'].values, df_teste['club_2'].values
clubs = np.unique(clubs)
clubs = np.delete(clubs,0)
clubes=[]
for i in range(len(clubs)):
    clubes.append([i+401, clubs[i], 'Colombia',1500])
df_clubcol = pd.DataFrame(clubes)

#cria o cadastro de times equatorianos 
df_teste = df_ecu[df_ecu['date']>='2002.01.01'][df_ecu['tournament']>='Liga']
clubs = df_teste['club_1'].values, df_teste['club_2'].values
clubs = np.unique(clubs)
clubes=[]
for i in range(len(clubs)):
    clubes.append([i+501, clubs[i], 'Equador',1500])
df_clubecu = pd.DataFrame(clubes)

#cria o cadastro de times paraguaios 
df_teste = df_par[df_par['date']>='2002.01.01'][df_par['tournament']=='Primera Division']
clubs = df_teste['club_1'].values, df_teste['club_2'].values
clubs = np.unique(clubs)
clubs = np.delete(clubs,3)
clubs = np.delete(clubs,3)
clubes=[]
for i in range(len(clubs)):
    clubes.append([i+601, clubs[i], 'Paraguai',1500])
df_clubpar = pd.DataFrame(clubes)

#cria o cadastro de times peruanos 
df_teste = df_per[df_per['date']>='2002.01.01'][df_per['tournament']>='Liga 1']
clubs = df_teste['club_1'].values, df_teste['club_2'].values
clubs = np.unique(clubs)
clubs = np.delete(clubs,1)
clubes=[]
for i in range(len(clubs)):
    clubes.append([i+701, clubs[i], 'Peru',1500])
df_clubper = pd.DataFrame(clubes)

#cria o cadastro de times uruguaios 
df_teste = df_uru[df_uru['date']>='2002.01.01'][df_uru['tournament']=='Primera Division']
clubs = df_teste['club_1'].values, df_teste['club_2'].values
clubs = np.unique(clubs)
clubs = np.delete(clubs,0)
clubes=[]
for i in range(len(clubs)):
    clubes.append([i+801, clubs[i], 'Uruguai',1500])
df_cluburu = pd.DataFrame(clubes)

#cria o cadastro de times venezuelanos 
df_teste = df_ven[df_ven['date']>='2002.01.01'][df_ven['tournament']=='Primera Division']
clubs = df_teste['club_1'].values, df_teste['club_2'].values
clubs = np.unique(clubs)
clubes=[]
for i in range(len(clubs)):
    clubes.append([i+901, clubs[i], 'Venezuela',1500])
df_clubven = pd.DataFrame(clubes)

#cria um dataframe para cadastrar os times
cols = ['id','nome','pais','rate']
join = [df_clubarg,df_clubbol,df_clubbra,df_clubchi,df_clubcol,df_clubecu,df_clubpar,df_clubper,df_cluburu,df_clubven]
df_teams = pd.concat(join)
df_teams.columns=cols
df_teams

#salva o cadastro num .csv
df_teams.to_csv('data/teams1500.csv')

Temos 341 times na base, conforme a tabela abaixo:

df_teams

Para que nosso modelo funcione, vamos precisar considerar somente os jogos entre times que estão na nossa base.

O primeiro passo será fazer esse filtro dentro da lista de jogos de cada país.


#PrepararJogos.py
    
cols = ['date','t1','g1','g2','t2','tourn']

#cria o dataframe de jogos argentinos
df_jogosarg = pd.DataFrame(columns = cols)
j=0
for i in range(df_arg['date'].size):
    jog = df_arg.iloc[[i]]
    if jog.iat[0,0]>='2002.01.01':
        if jog.iat[0,3]>=0 and jog.iat[0,6]>=0:
            jog.iat[0,2]=int(jog.iat[0,3])
            jog.iat[0,5]=int(jog.iat[0,6])    
        jogo = [jog.iat[0,0],jog.iat[0,1],jog.iat[0,2],jog.iat[0,5],jog.iat[0,4], jog.iat[0,7]]
        if jogo[1] in (df_teams[df_teams.pais == 'Argentina']['nome'].values):
            if jogo[4] in (df_teams[df_teams.pais == 'Argentina']['nome'].values):                   
                df_jogosarg.loc[j] = jogo
                j+=1

#exporta o dataframe para um .csv         
df_jogosarg.to_csv('data/jogosarg.csv')

#cria o dataframe de jogos bolivianos
df_jogosbol = pd.DataFrame(columns = cols)
j=0
for i in range(df_bol['date'].size):
    jog = df_bol.iloc[[i]]
    if jog.iat[0,0]>='2002.01.01':
        if jog.iat[0,3]>=0 and jog.iat[0,6]>=0:
            jog.iat[0,2]=int(jog.iat[0,3])
            jog.iat[0,5]=int(jog.iat[0,6])    
        jogo = [jog.iat[0,0],jog.iat[0,1],jog.iat[0,2],jog.iat[0,5],jog.iat[0,4], jog.iat[0,7]]
        if jogo[1] in (df_teams[df_teams.pais == 'Bolivia']['nome'].values):
            if jogo[4] in (df_teams[df_teams.pais == 'Bolivia']['nome'].values):                   
                df_jogosbol.loc[j] = jogo
                j+=1

#exporta o dataframe para um .csv  
df_jogosbol.to_csv('data/jogosbol.csv')


#cria o dataframe de jogos brasileiros
df_jogosbra = pd.DataFrame(columns = cols)
j=0
for i in range(df_bra['date'].size):
    jog = df_bra.iloc[[i]]
    if jog.iat[0,0]>='2002.01.01':
        if jog.iat[0,3]>=0 and jog.iat[0,6]>=0:
            jog.iat[0,2]=int(jog.iat[0,3])
            jog.iat[0,5]=int(jog.iat[0,6])    
        jogo = [jog.iat[0,0],jog.iat[0,1],jog.iat[0,2],jog.iat[0,5],jog.iat[0,4], jog.iat[0,7]]
        if jogo[1] in (df_teams[df_teams.pais == 'Brasil']['nome'].values):
            if jogo[4] in (df_teams[df_teams.pais == 'Brasil']['nome'].values):                   
                df_jogosbra.loc[j] = jogo
                j+=1

#exporta o dataframe para um .csv  
df_jogosbra.to_csv('data/jogosbra.csv')

#cria o dataframe de jogos chilenos
df_jogoschi = pd.DataFrame(columns = cols)
j=0
for i in range(df_chi['date'].size):
    jog = df_chi.iloc[[i]]
    if jog.iat[0,0]>='2002.01.01':
        if jog.iat[0,3]>=0 and jog.iat[0,6]>=0:
            jog.iat[0,2]=int(jog.iat[0,3])
            jog.iat[0,5]=int(jog.iat[0,6])    
        jogo = [jog.iat[0,0],jog.iat[0,1],jog.iat[0,2],jog.iat[0,5],jog.iat[0,4], jog.iat[0,7]]
        if jogo[1] in (df_teams[df_teams.pais == 'Chile']['nome'].values):
            if jogo[4] in (df_teams[df_teams.pais == 'Chile']['nome'].values):                   
                df_jogoschi.loc[j] = jogo
                j+=1
            
#exporta o dataframe para um .csv 
df_jogoschi.to_csv('data/jogoschi.csv')

#cria o dataframe de jogos colombianos
df_jogoscol = pd.DataFrame(columns = cols)
j=0
for i in range(df_col['date'].size):
    jog = df_col.iloc[[i]]
    if jog.iat[0,0]>='2002.01.01':
        if jog.iat[0,3]>=0 and jog.iat[0,6]>=0:
            jog.iat[0,2]=int(jog.iat[0,3])
            jog.iat[0,5]=int(jog.iat[0,6])    
        jogo = [jog.iat[0,0],jog.iat[0,1],jog.iat[0,2],jog.iat[0,5],jog.iat[0,4], jog.iat[0,7]]
        if jogo[1] in (df_teams[df_teams.pais == 'Colombia']['nome'].values):
            if jogo[4] in (df_teams[df_teams.pais == 'Colombia']['nome'].values):                   
                df_jogoscol.loc[j] = jogo
                j+=1

#exporta o dataframe para um .csv 
df_jogoscol.to_csv('data/jogoscol.csv')

#cria o dataframe de jogos equatorianos
df_jogosecu = pd.DataFrame(columns = cols)
j=0
for i in range(df_ecu['date'].size):
    jog = df_ecu.iloc[[i]]
    if jog.iat[0,0]>='2002.01.01':
        if jog.iat[0,3]>=0 and jog.iat[0,6]>=0:
            jog.iat[0,2]=int(jog.iat[0,3])
            jog.iat[0,5]=int(jog.iat[0,6])    
        jogo = [jog.iat[0,0],jog.iat[0,1],jog.iat[0,2],jog.iat[0,5],jog.iat[0,4], jog.iat[0,7]]
        if jogo[1] in (df_teams[df_teams.pais == 'Equador']['nome'].values):
            if jogo[4] in (df_teams[df_teams.pais == 'Equador']['nome'].values):                   
                df_jogosecu.loc[j] = jogo
                j+=1

#exporta o dataframe para um .csv 
df_jogosecu.to_csv('data/jogosecu.csv')

#cria o dataframe de jogos paraguaios
df_jogospar = pd.DataFrame(columns = cols)
j=0
for i in range(df_par['date'].size):
    jog = df_par.iloc[[i]]
    if jog.iat[0,0]>='2002.01.01':
        if jog.iat[0,3]>=0 and jog.iat[0,6]>=0:
            jog.iat[0,2]=int(jog.iat[0,3])
            jog.iat[0,5]=int(jog.iat[0,6])    
        jogo = [jog.iat[0,0],jog.iat[0,1],jog.iat[0,2],jog.iat[0,5],jog.iat[0,4], jog.iat[0,7]]
        if jogo[1] in (df_teams[df_teams.pais == 'Paraguai']['nome'].values):
            if jogo[4] in (df_teams[df_teams.pais == 'Paraguai']['nome'].values):                   
                df_jogospar.loc[j] = jogo
                j+=1

df_jogospar.to_csv('data/jogospar.csv')

#cria o dataframe de jogos peruanos
df_jogosper = pd.DataFrame(columns = cols)
j=0
for i in range(df_per['date'].size):
    jog = df_per.iloc[[i]]
    if jog.iat[0,0]>='2002.01.01':
        if jog.iat[0,3]>=0 and jog.iat[0,6]>=0:
            jog.iat[0,2]=int(jog.iat[0,3])
            jog.iat[0,5]=int(jog.iat[0,6])    
        jogo = [jog.iat[0,0],jog.iat[0,1],jog.iat[0,2],jog.iat[0,5],jog.iat[0,4], jog.iat[0,7]]
        if jogo[1] in (df_teams[df_teams.pais == 'Peru']['nome'].values):
            if jogo[4] in (df_teams[df_teams.pais == 'Peru']['nome'].values):                   
                df_jogosper.loc[j] = jogo
                j+=1

#exporta o dataframe para um .csv
df_jogosper.to_csv('data/jogosper.csv')

#cria o dataframe de jogos uruguaios
df_jogosuru = pd.DataFrame(columns = cols)
j=0
for i in range(df_uru['date'].size):
    jog = df_uru.iloc[[i]]
    if jog.iat[0,0]>='2002.01.01':
        if jog.iat[0,3]>=0 and jog.iat[0,6]>=0:
            jog.iat[0,2]=int(jog.iat[0,3])
            jog.iat[0,5]=int(jog.iat[0,6])    
        jogo = [jog.iat[0,0],jog.iat[0,1],jog.iat[0,2],jog.iat[0,5],jog.iat[0,4], jog.iat[0,7]]
        if jogo[1] in (df_teams[df_teams.pais == 'Uruguai']['nome'].values):
            if jogo[4] in (df_teams[df_teams.pais == 'Uruguai']['nome'].values):                   
                df_jogosuru.loc[j] = jogo
                j+=1

#exporta o dataframe para um .csv
df_jogosuru.to_csv('data/jogosuru.csv')

#cria o dataframe de jogos venezuelanos
df_jogosven = pd.DataFrame(columns = cols)
j=0
for i in range(df_ven['date'].size):
    jog = df_ven.iloc[[i]]
    if jog.iat[0,0]>='2002.01.01':
        if jog.iat[0,3]>=0 and jog.iat[0,6]>=0:
            jog.iat[0,2]=int(jog.iat[0,3])
            jog.iat[0,5]=int(jog.iat[0,6])    
        jogo = [jog.iat[0,0],jog.iat[0,1],jog.iat[0,2],jog.iat[0,5],jog.iat[0,4], jog.iat[0,7]]
        if jogo[1] in (df_teams[df_teams.pais == 'Venezuela']['nome'].values):
            if jogo[4] in (df_teams[df_teams.pais == 'Venezuela']['nome'].values):                   
                df_jogosven.loc[j] = jogo
                j+=1

#exporta o dataframe para um .csv
df_jogosven.to_csv('data/jogosven.csv')



Agora, faremos o mesmo para os jogos das competições continentais.

#PrepJogosSula.py

#lê o .csv e gera um dataframe com os jogos das competições continentais
df_sula = pd.read_csv("data/Continentais.csv")

df_teams = pd.read_csv('data/teams1500.csv')

dates = []
for i in range(df_sula['date'].size):        
    item = df_sula['date'].get(i)
    date = item.split('.')
    item = date[2]+'.'+date[1]+'.'+date[0]
    dates.append(item)    
df_sula['date']=dates
df_sula.sort_values(by="date", inplace=True)


cols = ['date','t1','g1','g2','t2','tourn']

#cria o dataframe de jogos das competições sul americanas
df_jogossula = pd.DataFrame(columns = cols)
j=0
for i in range(df_sula['date'].size):
    jog = df_sula.iloc[[i]]
    if jog.iat[0,0]>='2002.01.01': #descarta as partidas anteriores a 2002
        if jog.iat[0,3]>=0 and jog.iat[0,6]>=0: #ajeita o placar das partidas que foram decididas nos pênaltis
            jog.iat[0,2]=int(jog.iat[0,3])
            jog.iat[0,5]=int(jog.iat[0,6])    
        jogo = [jog.iat[0,0],jog.iat[0,1],jog.iat[0,2],jog.iat[0,5],jog.iat[0,4], jog.iat[0,7]]
        jogo[1] = re.sub("\s\[[A-Za-z]*\]", "", jogo[1])
        jogo[4] = re.sub("\s\[[A-Za-z]*\]", "", jogo[4])
        if jogo[1] in (df_teams['nome'].values): #descarta as partidas de times mexicanos ou que não jogaram primeira divisão desde 2002
            if jogo[4] in (df_teams['nome'].values):                   
                df_jogossula.loc[j] = jogo
                j+=1
                
#exporta o dataframe para um .csv
df_jogossula.to_csv('data/jogossula.csv')

Neste ponto, temos um problema.

Algumas das equipes nesta lista de jogos compartilham o mesmo nome. 

É o caso dos times na lista abaixo:

#encontra nomes duplicados de times na base
print(df_teams[df_teams.nome.duplicated(keep=False)])
nomes_duplicados = df_teams[df_teams.nome.duplicated()].nome.values
print('\n',nomes_duplicados)

Aqui, podemos ver quantas vezes esses nomes aparecem na nossa lista de jogos:

#verifica quantos jogos têm times com nome duplicados
print(df_jogossula[df_jogossula['t1'].isin(nomes_duplicados)]['t1'].size, 'mandantes entre os duplicados')
print(df_jogossula[df_jogossula['t2'].isin(nomes_duplicados)]['t2'].size, 'visitantes entre os duplicados')

Para resolver o problema, precisaremos listar os times pelo Id.

Primeiro, devemos identificar cada ocorrência dos times com nome repetido e trocá-lo pelo respectivo Id. Depois, faremos uma busca na base pelo Id dos times restantes.

Realizamos uma busca manual para identificar, dentre os times na base, aqueles que realmente disputaram jogos nas competições continentais.

Por sorte, ocorreram somente dois casos de times com o mesmo nome nessas competições: U. Catolica (Chile e Equador) e River Plate (Argentina e Uruguai).

Após a identificação manual, decidimos acrescentar um '2' ao final do nome para identificar os times com menos participações.

#identifica as partidas disputadas pela U. Catolica do Equador
print(df_jogossula[df_jogossula['t1']=='U. Catolica'][df_jogossula['tourn']=='Copa Sudamericana'][df_jogossula['date']>='2014']['date'])
for num in [2194,2249,2402,2633,2859,2898,3320,3397,3461]:
    df_jogossula.at[num,'t1']='U. Catolica2'

for num in [2210,2242,2418,2621,2763,2874,3375,3418,3446]:
    df_jogossula.at[num,'t2']='U. Catolica2'
print(df_jogossula[df_jogossula['t1']=='U. Catolica2'])
print(df_jogossula[df_jogossula['t2']=='U. Catolica2'])

#identifica as partidas disputadas pelo River Plate do Uruguai
for num in [1018,1215,1224,1241,1249,1393,1996,2029,2218,2253,2490,2500,2547,2558,3247,3399]:
    df_jogossula.at[num,'t1']='River Plate2'

for num in [1024,1199,1232,1244,1251,1383,1986,2018,2202,2245,2495,2523,2531,2572,3264,3416]:
    df_jogossula.at[num,'t2']='River Plate2'
print(df_jogossula[df_jogossula['t1']=='River Plate2'])
print(df_jogossula[df_jogossula['t2']=='River Plate2'])

Agora, estamos prontos para substituir o nome de cada time por seu respectivo Id, lembrando, é claro, de especificar o Id dos times com nome repetido na base.

#RemoveDuplicatas.py


#cria colunas para inserir os ids dos times em cada partida
mandantes = df_jogossula['t1']
visitantes = df_jogossula['t2']
idm = []
idv = []

#alimenta as colunas de id considerando os times duplicados:
for time in mandantes:
    if time=='Independiente':
        idm.append(22)
    elif time=='San Lorenzo':
        idm.append(33)
    elif time=='Santa Cruz':
        idm.append(235)
    elif time=='Guarani':
        idm.append(611)
    elif time=='Sport Boys':
        idm.append(125)
    elif time=='Portuguesa':
        idm.append(234)
    elif time=='River Plate':
        idm.append(31)
    elif time=='River Plate2':
        idm.append(826)
    elif time=='U. Catolica':
        idm.append(329)
    elif time=='U. Catolica2':
        idm.append(525)
    else:
        idm.append(int(df_teams[df_teams.nome==time].id.values))

for time in visitantes:
    if time=='Independiente':
        idv.append(22)
    elif time=='San Lorenzo':
        idv.append(33)
    elif time=='Santa Cruz':
        idv.append(235)
    elif time=='Guarani':
        idv.append(611)
    elif time=='Sport Boys':
        idv.append(125)
    elif time=='Portuguesa':
        idv.append(234)
    elif time=='River Plate':
        idv.append(31)
    elif time=='River Plate2':
        idv.append(826)
    elif time=='U. Catolica':
        idv.append(329)
    elif time=='U. Catolica2':
        idv.append(525)
    else:
        idv.append(int(df_teams[df_teams.nome==time].id.values))


#adiciona as colunas de ids ao dataframe
df_jogossula = df_jogossula.assign(id1=idm,id2=idv)[['date', 't1', 'id1', 'g1', 'g2',  'id2', 't2', 'tourn']]

#exporta o dataframe para um .csv
df_jogossula.to_csv('data/jogossulaid.csv')

Pronto!

Nossa base de jogos das competições continentais ficou assim:

df_jogossula

Agora, para uniformizar a formatação e facilitar o trabalho dos algoritmos do nosso modelo, faremos a identificação dos Ids dos times nas tabelas de jogos de cada país.

Vamos armazenar toda a base de jogos num único arquivo.

#CriaJogosId.py

#lê os .csv e gera um dataframe para cada país
df_jarg = pd.read_csv("data/jogosarg.csv")
df_jbol = pd.read_csv("data/jogosbol.csv")
df_jbra = pd.read_csv("data/jogosbra.csv")
df_jchi = pd.read_csv("data/jogoschi.csv")
df_jcol = pd.read_csv("data/jogoscol.csv")
df_jecu = pd.read_csv("data/jogosecu.csv")
df_jpar = pd.read_csv("data/jogospar.csv")
df_jper = pd.read_csv("data/jogosper.csv")
df_juru = pd.read_csv("data/jogosuru.csv")
df_jven = pd.read_csv("data/jogosven.csv")

df_jogosid = pd.read_csv("data/jogossulaid.csv")
df_teams = pd.read_csv('data/teams1500.csv')

dfs = [df_jarg,df_jbol,df_jbra,df_jchi,df_jcol,df_jecu,df_jpar,df_jper,df_juru,df_jven]


#cria colunas para inserir os ids dos times em cada partida

for i in range(len(dfs)):
    mandantes = dfs[i]['t1']
    visitantes = dfs[i]['t2']
    idm = []
    idv = []
    idlims = [100*i+j for j in range(100)]
    df_times=df_teams[df_teams["id"].isin(idlims)]
    
    #alimenta as colunas de id considerando os times duplicados:
    for time in mandantes:
        idm.append(int(df_times[df_times.nome==time].id.values))
    
    for time in visitantes:
        idv.append(int(df_times[df_times.nome==time].id.values))
    
    #adiciona as colunas de ids ao dataframe
    dfs[i] = dfs[i].assign(id1=idm,id2=idv)[['date', 't1', 'id1', 'g1', 'g2',  'id2', 't2', 'tourn']]
    df_jogosid = df_jogosid.append(dfs[i],ignore_index=True)
    print(i,dfs[i].index.size,df_jogosid.index.size)
    
df_jogosid = df_jogosid.sort_values(by=["date"])
df_jogosid = df_jogosid[['date', 't1', 'id1', 'g1', 'g2',  'id2', 't2', 'tourn']]
    
#exporta o dataframe para um .csv
df_jogosid.to_csv('data/jogosid.csv')

Voilá!

Temos nossa tabela de jogos, num total de 60900, com todos os times bem identificados na base!

df_jogosid

Chegamos então ao fim desta seção, e podemos passar à fase de exploração dos dados.
